requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
requests==2.32.3
python-dotenv==1.0.1
